requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
pandas>=2.2.0
streamlit>=1.30.0
python-dotenv>=1.0.0
fastapi>=0.108.0
uvicorn>=0.25.0
requests>=2.31.0